no-gui-desktop-player-python/requirements.txt
requests
mpv
python-socketio